Rename the logs directly to logging.
