+++ huxley/apps.py
@@ -1,0 +1,9 @@
+# Copyright (c) 2011-2015 Berkeley Model United Nations. All rights reserved.
+# Use of this source code is governed by a BSD License (see LICENSE).
+
+from django.apps import AppConfig
+
+
+class LoggingConfig(AppConfig):
+    name = 'huxley.logging'
+    label = 'logs'
