+++ huxley/logging/models.py
@@ -3,6 +3,7 @@
 
 from django.db import models
 
+
 class LogEntry(models.Model):
     level = models.CharField(max_length=200)
     message = models.TextField()
